+++ app/Modules/Prescriptions/Requests/UpdatePrescriptionRequest.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace App\Modules\Prescriptions\Requests;
+
+class UpdatePrescriptionRequest extends StorePrescriptionRequest {}